app/requirements.txt
streamlit==1.22.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pillow==9.5.0
psycopg2-binary==2.9.6